pyproject.toml
[project]
name = "rainier"
version = "0.2.0"
description = "Trading analysis platform — futures price action + stock money flow"
requires-python = ">=3.12"
dependencies = [
    # Core
    "click>=8.1",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    # Data
    "pandas>=2.2",
    "numpy>=1.26",
    "yfinance>=1.2.0",
    # Database
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    # Visualization
    "plotly>=5.18",
    "kaleido==0.2.1",
    # Web scraping
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
    # Scheduling
    "apscheduler>=3.10",
    # LLM
    "litellm>=1.40",
    # Notifications
    "httpx>=0.27",
    "apprise>=1.9.8",
    # Logging
    "structlog>=24.0",
]

[project.scripts]
rainier = "rainier.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rainier"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[project.optional-dependencies]
dashboard = ["streamlit>=1.30"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
]
